pyproject.toml
[project]
name = "cognito-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.66",
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.4.0",
    "slowapi>=0.1.9",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0",
    "uvicorn>=0.41.0",
]
